+++ dune/typetree/treepath.hh
@@ -492,6 +492,36 @@ namespace Dune {
       return std::bool_constant<hybridTreePath(lhs...) != hybridTreePath(rhs...)>{};
     }
 
+#if __cpp_impl_three_way_comparison >= 201907
+
+    //! Compare two `HybridTreePath`s ordering
+    template<class... S, class... T>
+    requires (sizeof...(S) == sizeof...(T))
+    constexpr auto operator<=>(const HybridTreePath<S...>& lhs, const HybridTreePath<T...>& rhs)
+    {
+      return (lhs._data <=> rhs._data);
+    }
+
+    //! Compare two `HybridTreePath`s ordering
+    template<class... S, class... T>
+    requires (sizeof...(S) != sizeof...(T))
+    constexpr auto operator<=>(const HybridTreePath<S...>& lhs, const HybridTreePath<T...>& rhs)
+    {
+      // utility to make a common sized tuple
+      auto is = std::make_index_sequence<std::min(sizeof...(S), sizeof...(T))>{};
+      auto make_common_tuple = [is](const auto& tuple){
+        return [&]<std::size_t... i>(std::index_sequence<i...>){
+          return std::make_tuple(std::get<i>(tuple)...);
+        }(is);
+      };
+      // three way comparison between equally sized tuples
+      const auto comp = make_common_tuple(lhs._data) <=> make_common_tuple(rhs._data);
+      if (comp != 0) return comp;
+      // if the common part is equal, decide by the size
+      return sizeof...(S) <=> sizeof...(T);
+    }
+#endif // __cpp_impl_three_way_comparison
+
 
     inline namespace Literals {